+++ Python/motor.py
@@ -236,11 +236,11 @@ class Motor(AvionicsObject):
         dialog.setWindowModality(Qt.ApplicationModal)
 
         # Set dialog size and place in middle of window
-        dialog.resize(450 * self.gui.pixel_scale_ratio[0], 240 * self.gui.pixel_scale_ratio[1])
-        dialog.setMinimumWidth(450 * self.gui.pixel_scale_ratio[0])
-        dialog.setMinimumWidth(240 * self.gui.pixel_scale_ratio[1])
-        dialog.move((self.window.width() - dialog.width()) / 2,
-                    (self.window.height() - dialog.height()) / 2)
+        dialog.resize(int(450 * self.gui.pixel_scale_ratio[0]), int(240 * self.gui.pixel_scale_ratio[1]))
+        dialog.setMinimumWidth(int(450 * self.gui.pixel_scale_ratio[0]))
+        dialog.setMinimumWidth(int(240 * self.gui.pixel_scale_ratio[1]))
+        dialog.move(int((self.window.width() - dialog.width()) / 2),
+                    int((self.window.height() - dialog.height()) / 2))
 
         # Vertical layout to hold everything
         verticalLayout = QVBoxLayout(dialog)
@@ -324,14 +324,14 @@ class Motor(AvionicsObject):
         cancel_button.setDefault(False)
         cancel_button.setAutoDefault(False)
         cancel_button.clicked.connect(lambda: dialog.done(1))
-        cancel_button.setFixedWidth(125 * self.gui.pixel_scale_ratio[0])  # Lazy way to make buttons not full width
+        cancel_button.setFixedWidth(int(125 * self.gui.pixel_scale_ratio[0]))  # Lazy way to make buttons not full width
 
         save_button = QPushButton("Save")
         save_button.setFont(font)
         save_button.setDefault(False)
         save_button.setAutoDefault(False)
         save_button.clicked.connect(lambda: self.motorDialogSave(spinBoxes, dialog))
-        save_button.setFixedWidth(125 * self.gui.pixel_scale_ratio[0])
+        save_button.setFixedWidth(int(125 * self.gui.pixel_scale_ratio[0]))
 
         buttonLayout.addWidget(cancel_button)
         buttonLayout.addWidget(save_button)
